+++ Future/src/Platform/Windows/WindowsWindow.cpp
@@ -4,6 +4,7 @@
 #include "Future/Events/ApplicationEvent.h"
 #include "Future/Events/KeyEvent.h"
 #include "Future/Events/MouseEvent.h"
+#include "Platform/OpenGL/OpenGLContext.h"
 
 #include "glad/glad.h"
 
@@ -50,6 +51,9 @@ namespace Future
 
 		// create GLFWwindow and make it current context
 		m_Window = glfwCreateWindow((int)m_Data.Width, (int)m_Data.Height, m_Data.Title.c_str(), nullptr, nullptr);
+
+		
+
 		glfwMakeContextCurrent(m_Window);
 
 		// initialize Glad
@@ -130,7 +134,7 @@ namespace Future
 		{
 			WindowData* windowData = (WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseScrolledEvent event(xOffset, yOffset);
+			MouseScrolledEvent event((float) xOffset, (float) yOffset);
 			windowData->EventCallback(event);
 		});
 
@@ -138,7 +142,7 @@ namespace Future
 		{
 			WindowData* windowData = (WindowData*)glfwGetWindowUserPointer(window);
 			
-			MouseMovedEvent event(xPos, yPos);
+			MouseMovedEvent event((float) xPos, (float) yPos);
 			windowData->EventCallback(event);
 		});
